NAS-115487: Fixed naming
