+++ src/app/pages/sharing/components/shares-dashboard/shares-dashboard.component.ts
@@ -29,7 +29,6 @@ import { EntityUtils } from 'app/modules/entity/utils';
 import { TargetFormComponent } from 'app/pages/sharing/iscsi/target/target-form/target-form.component';
 import { NfsFormComponent } from 'app/pages/sharing/nfs/nfs-form/nfs-form.component';
 import { SmbFormComponent } from 'app/pages/sharing/smb/smb-form/smb-form.component';
-import { SmbForm2Component } from 'app/pages/sharing/smb/smb-form2/smb-form2.component';
 import { WebdavFormComponent } from 'app/pages/sharing/webdav/webdav-form/webdav-form.component';
 import {
   DialogService,
@@ -338,7 +337,6 @@ export class SharesDashboardComponent implements AfterViewInit {
     let formComponent: Type<
     NfsFormComponent |
     SmbFormComponent |
-    SmbForm2Component |
     WebdavFormComponent |
     TargetFormComponent
     >;
@@ -347,7 +345,7 @@ export class SharesDashboardComponent implements AfterViewInit {
         formComponent = NfsFormComponent;
         break;
       case ShareType.Smb:
-        formComponent = SmbForm2Component;
+        formComponent = SmbFormComponent;
         break;
       case ShareType.WebDav:
         formComponent = WebdavFormComponent;
@@ -365,7 +363,7 @@ export class SharesDashboardComponent implements AfterViewInit {
         } else if (share === ShareType.Nfs) {
           (form as NfsFormComponent).setNfsShareForEdit(row);
         } else if (share === ShareType.Smb) {
-          (form as SmbForm2Component).setSmbShareForEdit(row);
+          (form as SmbFormComponent).setSmbShareForEdit(row);
         }
       }
       this.slideInService.onClose$.pipe(untilDestroyed(this)).subscribe(() => {
