+++ src/app/pages/sharing/smb/smb-form/smb-form.component.scss
